BC-633 - refactor rooms types
